+++ clients/cxx/fileio/memory_fileio.cc
@@ -1,3 +1,16 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// see the license for the specific language governing permissions and
+// limitations under the license.
 #include "clients/cxx/fileio/memory_fileio.h"
 
 #include <memory>
@@ -109,7 +122,8 @@ bool FileIO::Read(std::string* serialized_record) {
     SetError("File cannot be read from. It has been deleted.");
     return false;
   }
-  if (read_idx_ >= file->size()) {
+  if (static_cast<std::size_t>(read_idx_) >=
+      file->size()) {
     SetError("EOF");
     read_eof_ = true;
     return false;
